+++ gDRutils/inst/CHANGELOG.md
@@ -1,3 +1,6 @@
+# Change to v.1.0.28
+- refactor identifier `drugname` to `drug_name`
+
 # Change to v.1.0.27
 - add `MAEpply` function